Related artists crawler endpoint + dump to MySQL
 * Create admin/management endpoints for the HTTP API to dump related artists to a MySQL table and crawl related artists for some random artists in the Redis hash set
 * Several finishing touches/polish items on the frontend
   * Add share links to the header on user stats pages that auto-populate with the current URL
   * Update styling on the compare to landing page by adding header, tweaking alignment/padding, etc.

+++ backend/src/spotify_api.rs
@@ -288,7 +288,7 @@ pub(crate) fn store_stats_snapshot(
                         mapped_spotify_id: mapped_artist_spotify_ids[&artist_spotify_id],
                         update_time,
                         timeframe: map_timeframe_to_timeframe_id(&artist_timeframe),
-                        ranking: artist_ranking as u16,
+                        ranking: artist_ranking as u8,
                     }
                 })
         })
@@ -375,7 +375,7 @@ pub(crate) fn store_stats_snapshot(
                         mapped_spotify_id: mapped_track_spotify_ids[&track_spotify_id],
                         update_time,
                         timeframe: map_timeframe_to_timeframe_id(&track_timeframe),
-                        ranking: track_ranking as u16,
+                        ranking: track_ranking as u8,
                     },
                 )
         })
@@ -605,6 +605,7 @@ pub(crate) fn get_related_artists(
     Ok(res.artists)
 }
 
+/// `artist_ids` must not have any duplicates
 pub(crate) fn get_multiple_related_artists(
     bearer_token: String,
     artist_ids: &[&str],
